Refactor access modifiers and add documentation
Changed access modifiers of interfaces and classes from
`internal` to `public` for broader accessibility. Added
XML documentation comments to clarify the purpose and
functionality of methods and properties. Overall structure
and functionality remain unchanged, enhancing visibility
and documentation for the bowling search functionality.
