--- source/main/Bowlers/Search/ISearchBowlersView.cs
@@ -1,12 +1,0 @@
-﻿
-namespace BowlingMegabucks.TournamentManager.Bowlers.Search;
-internal interface IView
-{
-    void BindResults(IEnumerable<IViewModel> bowlers);
-
-    void DisplayError(string message);
-
-    void DisplayMessage(string message);
-
-    Models.BowlerSearchCriteria SearchCriteria { get; }
-}
